+++ .credo.exs
@@ -56,7 +56,7 @@
         {Credo.Check.Refactor.MapJoin, []},
         {Credo.Check.Refactor.NegatedConditionsInUnless, []},
         {Credo.Check.Refactor.NegatedConditionsWithElse, []},
-        {Credo.Check.Refactor.Nesting, [max_nesting: 5]},
+        {Credo.Check.Refactor.Nesting, [max_nesting: 6]},
         {Credo.Check.Refactor.RedundantWithClauseResult, []},
         {Credo.Check.Refactor.RejectFilter, []},
         {Credo.Check.Refactor.UnlessWithElse, []},